upd: reactivate test case that is skipped due to failed assumption
When this test case was introduced. The semantics of "getById"-Methods
where slightly different. These methods would throw an
IllegalArgumentException when the Object was not found. This behavior
was changed to returning "null" when the Object is not found. When this
change occurred this testcase was missed.

The general intended logic is retained: Guess random ids until one
unused is found (now checking for null and not for an exception). If 20
guesses fail (which is nearly impossible) the test case is skipped
instead of failed.

Change-Id: Id328d3c068f375488862bd06cfa0daf42a8c425f