add: handling of RA Agent Contract
Only covers the basic functionality.

The full text of the contract and the email will be covered in a later
patch.

The data contract.id and contract.token will be used for pdf-output in a
later patch.

The implementation of the restrictions connected to signed contract will
be covered in a later patch.

Change-Id: I5b47d31458779d227a4f9702a9e7563ab210e7e5